Finish creating Event Creation page #47
Finished the implementation of the event creation page. Saves input from users as states, and even checks if all the fields have been filled out. Imported the library mentioned in CreateEvent.js:
https://github.com/wojtekmaj/react-date-picker

Also added a button that can be used to get to the event creation page that's only showed in the feed page. It's called PostButton.js