Create Demo for Sess Auth. #56
This php file detects if the browser has a session id saved from a prior browsing session in the browser's cookies. If not, it creates a new session id, puts it into a persistent cookie, and inserts it into the sessions table in our database.

Our sessions table has an event that checks for expired tokens and deletes them if they are.