Create session verifier & enforcehttps #63
I created new php files titled verify-session that checks if there is a persistent session cookie that exists and verifies if it's a correct one.

Global functions enforceHTTPS and checkSessionId created in app.js. Ideally, put these in useEffect so they are executed on render.

Create-event.php has been created, but no progress