Create event filter menu #85
Worked on the front end of things. Clicking on "Filter" brings up checkboxes. Checking a box means that you want to see those types of events. Not selecting anything will mean all types of events are allowed to be seen in feed.

Clicking the filter posts button in the future should cause the updatePage variable to change, triggering the createEffect (and thus axios) which will repopulate the page with the filtered events