Make event code verifier #96
Started making progress for this feature. Spent around 4 hours adding the event code to create event form and testing event code submission. If a user has already submitted a correct code for an event, then they won't add anymore rows to the data table. If the user hasn't done it for a specific event yet and submitted a wrong code, it will notify them.

User id is hard coded, have to do it based on cookies later