Finish search event feature #103
Had to alter FeedArea.js to include 2 new props: 1 that determines if the filter button is shown and 2 is for a query created by the search event feature. In the home feed, the filter is shown and an empty string is passed for query prop. For search feature, filter not shown and query prop gets updated with help of 2 state variables.

Added the search link to mobile view and the site name redirects to feed now.

PostButton got moved to app.js from feedarea so not shown in search-event.

Leaderboard.css needs fixing